% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CalculateBins.R
\name{CalculateBins}
\alias{CalculateBins}
\title{Internal function for creating default bins for dataset variables.}
\usage{
CalculateBins(dsObserved, binCountSuggestion = 30L)
}
\arguments{
\item{dsObserved}{The \code{data.frame} with columns to calculate bins.}

\item{binCountSuggestion}{An \code{integer} or \code{numeric} value for the suggested number of bins for each variable.}
}
\value{
Returns a \code{list}, with two elements.  Each element is an array with as many values as columns in \code{dsObserved}.
\enumerate{
\item{\code{binWidth} The variable name (in \code{dsObserved}).}
\item{\code{binStart} The variable's \code{\link{class}}. (eg, numeric, Date, factor)}
}
}
\description{
An internal function (ie, that's not currently exposed/exported outside the package) for creating default bins for dataset variables.
}
\examples{
#TabularManifest:::CalculateBins(dsObserved=datasets::freeny)
#TabularManifest:::CalculateBins(dsObserved=datasets::InsectSprays)
}

