% Generated by roxygen2 (4.0.0): do not edit by hand
\name{construct_graph}
\alias{construct_graph}
\alias{construct_graph_list_univariate}
\alias{construct_graph_univariate}
\title{Construct a graph or list of graphs}
\usage{
construct_graph_univariate(variableName, ds_metadata, dsObserved)
}
\arguments{
  \item{variableName}{The name of the single variable to
  graph.}

  \item{ds_metadata}{The \code{data.frame} containing the
  metadata. See
  \code{\link{CreateManifestExploreUnivariate}}.}

  \item{dsObserved}{The \code{data.frame} containing the
  data to be graphed.}
}
\description{
Construct a graph or list of graphs, whose characteristics are determined by a configuration file.
}
\examples{
#dsObserved <- beaver1
dsObserved <- InsectSprays
ds_manifest <- TabularManifest::CreateManifestExploreUnivariate(dsObserved, writeToDisk=FALSE)

construct_graph_univariate(variableName="count", ds_manifest, InsectSprays)

construct_graph_list_univariate(ds_manifest=ds_manifest, dsObserved=dsObserved)
}

