% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ConstructGraph.R
\name{ConstructGraph}
\alias{ConstructGraph}
\alias{ConstructGraphListUnivariate}
\alias{ConstructGraphUnivariate}
\title{Construct a graph or list of graphs}
\usage{
ConstructGraphUnivariate(variableName, dsMetadata, dsObserved)
}
\arguments{
\item{variableName}{The name of the single variable to graph.}

\item{dsMetadata}{The \code{data.frame} containing the metadata. See \code{\link{CreateManifestExploreUnivariate}}.}

\item{dsObserved}{The \code{data.frame} containing the data to be graphed.}
}
\description{
Construct a graph or list of graphs, whose characteristics are determined by a configuration file.
}
\examples{
#dsObserved <- beaver1
dsObserved <- InsectSprays
dsManifest <- TabularManifest::CreateManifestExploreUnivariate(dsObserved, writeToDisk=FALSE)

ConstructGraphUnivariate(variableName="count", dsManifest, InsectSprays)

ConstructGraphListUnivariate(dsManifest=dsManifest, dsObserved=dsObserved)
}

