% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-date.R
\name{histogram_date}
\alias{histogram_date}
\title{Generate a Histogram for a \code{date} variable.}
\usage{
histogram_date(
  d_observed,
  variable_name,
  bin_unit = c("day", "week", "month", "quarter", "year"),
  main_title = base::gsub("_", " ", variable_name, perl = TRUE),
  sub_title = NULL,
  caption = paste0("each bin is 1 ", bin_unit, " wide"),
  x_title = variable_name,
  y_title = "Frequency",
  x_axis_format = scales::comma_format(),
  font_base_size = 12
)
}
\arguments{
\item{d_observed}{The \code{data.frame} with the variable to graph.}

\item{variable_name}{The name of the variable to graph. \code{character}.}

\item{bin_unit}{The width of the histogram bins. Balue is passed to \code{\link[base:seq.Date]{base::seq.Date()}}. Defaults to 'day'. \code{numeric}.}

\item{main_title}{The desired title on top of the graph.  Defaults to \code{variable_name}, with underscores replaced with spaces. If no title is desired, pass a value of \code{NULL}. \code{character}.}

\item{sub_title}{The desired subtitle near the top of the graph.  Defaults to \code{NULL} If no subtitle is desired, pass a value of \code{NULL}. \code{character}.}

\item{caption}{The desired text in the bottom-right, below the axis.  Defaults to the \code{bin_width}. If no caption is desired, pass a value of \code{NULL}. \code{character}.}

\item{x_title}{The desired title on the \emph{x}-axis.  Defaults to the \code{variable_name}.  If no axis title is desired, pass a value of \code{NULL}. \code{character}.}

\item{y_title}{The desired title on the \emph{y}-axis.  Defaults to ``Frequency''. If no axis title is desired, pass a value of \code{NULL}. \code{character}.}

\item{x_axis_format}{How the \emph{x}-axis digits are formatted. Defaults to \code{scales::comma_format()}.  \code{scale format}.}

\item{font_base_size}{Sets font size through ggplot2's theme.}
}
\value{
Returns a histogram as a \code{ggplot2} object.
}
\description{
Generate a histogram for a \code{date} variable.  This graph is intended to quickly provide
the researcher with a quick, yet thorough representation of the date.  The additional annotations may not
be desired for publication-quality plots.
}
\examples{
# Don't run graphs on a headless machine without any the basic graphics packages installed.
if( require(grDevices) & require(nycflights13) ) {
  ds               <- nycflights13::flights
  ds$date_depart   <- as.Date(ISOdate(ds$year, ds$month, ds$day))
  ds$date_blank    <- as.Date(NA)
  
  histogram_date(ds, variable_name="date_depart", bin_unit="day")
  histogram_date(ds, variable_name="date_depart", bin_unit="week")
  histogram_date(ds, variable_name="date_depart", bin_unit="month")
  histogram_date(ds, variable_name="date_depart", bin_unit="quarter")
  histogram_date(ds, variable_name="date_depart", bin_unit="year")
  
  histogram_date(ds, variable_name="date_depart", bin_unit="day")
}
}
