% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistogramContinuous.R
\name{HistogramContinuous}
\alias{HistogramContinuous}
\title{Generate a Histogram for a \code{numeric} or \code{integer} variable.}
\usage{
HistogramContinuous(dsObserved, variableName, binWidth = NULL,
  mainTitle = variableName, xTitle = paste0(variableName, " (each bin is ",
  scales::comma(binWidth), " units wide)"), yTitle = "Frequency",
  roundedDigits = 0L)
}
\arguments{
\item{dsObserved}{The \code{data.frame} with the variable to graph.}

\item{variableName}{The name of the variable to graph. \code{character}.}

\item{binWidth}{The width of the histogram bins. If NULL, the \code{ggplot2} default is used. \code{numeric}.}

\item{mainTitle}{The desired title on top of the graph.  Defaults to \code{variableName}. If no title is desired, pass a value of \code{NULL}. \code{character}.}

\item{xTitle}{The desired title on the \emph{x}-axis.  Defaults to the \code{variableName} and the \code{binWidth}. If no axis title is desired, pass a value of \code{NULL}. \code{character}.}

\item{yTitle}{The desired title on the \emph{y}-axis.  Defaults to ``Frequency''. If no axis title is desired, pass a value of \code{NULL}. \code{character}.}

\item{roundedDigits}{The number of decimals to show for the mean and median annotations. \code{character}.}
}
\value{
Returns a histogram as a \code{ggplot2} object.
}
\description{
Generate a histogram for a \code{numeric} or \code{integer} variable.  This graph is intended to quickly provide
the researcher with a quick, yet thorough representation of the continuous variable.  The additional annotations may not
be desired for publication-quality plots.
}
\examples{
library(datasets)
#Don't run graphs on a headless machine without any the basic graphics packages installed.
if( require(grDevices) ) { 
  HistogramContinuous(dsObserved=beaver1, variableName="temp", binWidth=.1)
}
}

