% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-discrete.R
\name{HistogramDiscrete}
\alias{HistogramDiscrete}
\title{Generate a Histogram for a \code{character} or \code{factor} variable.}
\usage{
HistogramDiscrete(ds_observed, variable_name, levelsToExclude = character(0),
  mainTitle = variable_name, xTitle = NULL,
  yTitle = "Number of Included Records", textSizePercentage = 6,
  bin_width = 1L)
}
\arguments{
\item{ds_observed}{The \code{data.frame} with the variable to graph.}

\item{variable_name}{The name of the variable to graph. \code{character}.}

\item{levelsToExclude}{An array of of the levels to be excluded from the histogram. Pass an empty variable (\emph{ie}, \code{character(0)}) if all levels are desired; this is the default. \code{character}.}

\item{mainTitle}{The desired title on top of the graph.  Defaults to \code{variable_name}. If no title is desired, pass a value of \code{NULL}. \code{character}.}

\item{xTitle}{The desired title on the \emph{x}-axis.  Defaults to the number of included records. If no axis title is desired, pass a value of \code{NULL}. \code{character}.}

\item{yTitle}{The desired title on the \emph{y}-axis.  Defaults to ``Frequency''. If no axis title is desired, pass a value of \code{NULL}. \code{character}.}

\item{textSizePercentage}{The size of the percentage values on top of the bars. \code{character}.}

\item{bin_width}{(This parameter is included for compatibility with other graphing functions.  It should always be \code{1} for discrete and boolean variables.)}
}
\value{
Returns a histogram as a \code{ggplot2} object.
}
\description{
Generate a histogram for a \code{character} or \code{factor} variable.  This graph is intended to quickly provide
the researcher with a quick, yet thorough representation of the continuous variable.  The additional annotations may not
be desired for publication-quality plots.
}
\examples{
library(datasets)
#Don't run graphs on a headless machine without any the basic graphics packages installed.
if( require(grDevices) ) { 
  HistogramDiscrete(ds_observed=infert, variable_name="education")
  HistogramDiscrete(ds_observed=infert, variable_name="age")
}
}

