% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-model-continuous-x-binary-y-logit.R
\name{ScatterModelContinuousXBinaryYLogit}
\alias{ScatterModelContinuousXBinaryYLogit}
\title{Internal function for examining a logit performance}
\usage{
ScatterModelContinuousXBinaryYLogit(ds_plot, x_name, y_name = "y",
  yhat_name = "yhat", residual_name = "residual", alpha_point = 0.05,
  alpha_se_band = 0.15, xLabelFormat = scales::comma,
  color_smooth_observed = "#1b9e77", color_smooth_predicted = "#d95f02",
  color_smooth_residual = "#7570b3", vertical_limits = c(-0.05, 1.05),
  jitter_observed = ggplot2::position_jitter(w = 0, h = 0.2),
  jitter_predicted = ggplot2::position_jitter(w = 0, h = 0),
  seed_value = NA_real_)
}
\arguments{
\item{ds_plot}{The \code{data.frame} of observed and predicted values to plot.}

\item{x_name}{The name of the predictor \code{character}.}

\item{y_name}{The name of the observed response \code{character}.}

\item{yhat_name}{The name of the predicted response \code{character}.}

\item{residual_name}{The name of the model residual. \code{character}.}

\item{alpha_point}{The transparency of each plotted point. A \code{numeric} value from 0 to 1.}

\item{alpha_se_band}{The transparency of the standard error bands. A \code{numeric} value from 0 to 1.}

\item{xLabelFormat}{The name of the function used to format the \emph{x}-axis. \code{character}.}

\item{color_smooth_observed}{The plotted color of the observed values' GAM trend.  \code{character}.}

\item{color_smooth_predicted}{The plotted color of the predicted's GAM trend.  \code{character}.}

\item{color_smooth_residual}{The plotted color of the residual's GAM trend.  \code{character}.}

\item{vertical_limits}{The plotted limits of the response variable. A two-element \code{numeric} array.}

\item{jitter_observed}{A function dictating how the observed values are jittered.}

\item{jitter_predicted}{A function dictating how the predicted values are jittered.}

\item{seed_value}{The value of the RNG seed, which affects jittering. No seed is set if a value of \code{NA} is passed.  \code{numeric}.}
}
\description{
Internal function for examining a logit performance
}

