% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ScatterModelDiscreteXBinaryYLogit}
\alias{ScatterModelDiscreteXBinaryYLogit}
\title{Internal function for examining a logit performance}
\usage{
ScatterModelDiscreteXBinaryYLogit(dsPlot, xName, yName = "y",
  yHatName = "yhat", residualName = "residual", alphaPoint = 0.05,
  alphaSEBand = 0.15, xLabelFormat = scales::comma,
  colorSmoothObserved = "#1b9e77", colorSmoothPredicted = "#d95f02",
  colorSmoothResidual = "#7570b3", colorGroupCount = "tomato",
  verticalLimits = c(-0.05, 1.05),
  jitterObserved = ggplot2::position_jitter(w = 0.35, h = 0.2),
  jitterPredicted = ggplot2::position_jitter(w = 0.35, h = 0),
  seedValue = NA_real_)
}
\arguments{
  \item{dsPlot}{The \code{data.frame} of observed and
  predicted values to plot.}

  \item{xName}{The name of the predictor \code{character}.}

  \item{yName}{The name of the observed response
  \code{character}.}

  \item{yHatName}{The name of the predicted response
  \code{character}.}

  \item{residualName}{The name of the model residual.
  \code{character}.}

  \item{alphaPoint}{The transparency of each plotted point.
  A \code{numeric} value from 0 to 1.}

  \item{alphaSEBand}{The transparency of the standard error
  bands. A \code{numeric} value from 0 to 1.}

  \item{xLabelFormat}{The name of the function used to
  format the \emph{x}-axis. \code{character}.}

  \item{colorSmoothObserved}{The plotted color of the
  observed values' GAM trend.  \code{character}.}

  \item{colorSmoothPredicted}{The plotted color of the
  predicted's GAM trend.  \code{character}.}

  \item{colorSmoothResidual}{The plotted color of the
  residual's GAM trend.  \code{character}.}

  \item{colorGroupCount}{The color indicating how many
  cases belong to each level.  \code{character}.}

  \item{verticalLimits}{The plotted limits of the response
  variable. A two-element \code{numeric} array.}

  \item{jitterObserved}{A function dictating how the
  observed values are jittered.}

  \item{jitterPredicted}{A function dictating how the
  predicted values are jittered.}

  \item{seedValue}{The value of the RNG seed, which affects
  jittering. No seed is set if a value of \code{NA} is
  passed.  \code{numeric}.}
}
\description{
Internal function for examining a logit performance
}

