% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-trim.R
\name{univariate_trim}
\alias{univariate_trim}
\title{Trim extreme quantiles from a variable.}
\usage{
univariate_trim(scores, quantile_lower = 0, quantile_upper = 1)
}
\arguments{
\item{scores}{A vector of values that \code{quantile()} can accept.}

\item{quantile_lower}{The lower inclusive bound of values to retain.}

\item{quantile_upper}{The upper inclusive bound of values to retain.}
}
\value{
Returns a vector of the same \code{class} as \code{scores}.
}
\description{
Replaces extreme values with \code{NA} values.
}
\examples{
#Exclude the largest 5\%
TabularManifest::univariate_trim(scores=datasets::beaver1$temp,
  quantile_lower=0, quantile_upper=.95)

#Exclude the largest 5\% and the smallest 5\%
TabularManifest::univariate_trim(scores=datasets::beaver1$temp,
  quantile_lower=.05, quantile_upper=.95)
}
