% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/UnivariateTrim.R
\name{UnivariateTrim}
\alias{UnivariateTrim}
\title{Trim extreme quantiles from a variable.}
\usage{
UnivariateTrim(scores, quantileLower = 0, quantileUpper = 1)
}
\arguments{
\item{scores}{A vector of values that \code{quantile()} can accept.}

\item{quantileLower}{The lower inclusive bound of values to retain.}

\item{quantileUpper}{The upper inclusive bound of values to retain.}
}
\value{
Returns a vector of the same \code{class} as \code{scores}.
}
\description{
Replaces extreme values with \code{NA} values.
}
\examples{
#Exclude the largest 5\%
TabularManifest::UnivariateTrim(scores=datasets::beaver1$temp, quantileLower=0, quantileUpper=.95)

#Exclude the largest 5\% and the smallest 5\%
TabularManifest::UnivariateTrim(scores=datasets::beaver1$temp, quantileLower=.05, quantileUpper=.95)
}

