% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalculateRoundingDigits.R
\name{CalculateRoundingDigits}
\alias{CalculateRoundingDigits}
\title{Internal function for calculating rounding digits for dataset variables}
\usage{
CalculateRoundingDigits(dsObserved)
}
\arguments{
\item{dsObserved}{The \code{data.frame} with columns to calculate bins.}
}
\value{
Returns a \code{numeric} vector, indicating how many rounding digits *might* be appropriate.
  Each element is an array with as many values as columns in \code{dsObserved}.
}
\description{
An internal function (ie, that's not currently exposed/exported outside the package) for creating default bins for dataset variables.
}
\examples{
TabularManifest:::CalculateRoundingDigits(dsObserved=freeny)
TabularManifest:::CalculateRoundingDigits(dsObserved=InsectSprays)
TabularManifest:::CalculateRoundingDigits(dsObserved=beaver1)
}

