% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-graph.R
\name{construct_graph}
\alias{construct_graph}
\alias{construct_graph_univariate}
\alias{construct_graph_list_univariate}
\title{Construct a graph or list of graphs}
\usage{
construct_graph_univariate(variable_name, ds_metadata, d_observed)
}
\arguments{
\item{variable_name}{The name of the single variable to graph.}

\item{ds_metadata}{The \code{data.frame} containing the metadata. See \code{\link{create_manifest_explore_univariate}}.}

\item{d_observed}{The \code{data.frame} containing the data to be graphed.}
}
\description{
Construct a graph or list of graphs, whose characteristics are determined by a configuration file.
}
\examples{
#d_observed <- beaver1
d_observed <- InsectSprays
ds_manifest <- TabularManifest::create_manifest_explore_univariate(d_observed, write_to_disk=FALSE)

construct_graph_univariate(variable_name="count", ds_manifest, InsectSprays)

construct_graph_list_univariate(ds_manifest=ds_manifest, d_observed=d_observed)
}
