% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-graph.R
\name{construct_graph}
\alias{construct_graph}
\alias{construct_graph_list_univariate}
\alias{construct_graph_univariate}
\title{Construct a graph or list of graphs}
\usage{
construct_graph_univariate(variable_name, ds_metadata, ds_observed)
}
\arguments{
\item{variable_name}{The name of the single variable to graph.}

\item{ds_metadata}{The \code{data.frame} containing the metadata. See \code{\link{CreateManifestExploreUnivariate}}.}

\item{ds_observed}{The \code{data.frame} containing the data to be graphed.}
}
\description{
Construct a graph or list of graphs, whose characteristics are determined by a configuration file.
}
\examples{
#ds_observed <- beaver1
ds_observed <- InsectSprays
ds_manifest <- TabularManifest::CreateManifestExploreUnivariate(ds_observed, writeToDisk=FALSE)

construct_graph_univariate(variable_name="count", ds_manifest, InsectSprays)

construct_graph_list_univariate(ds_manifest=ds_manifest, ds_observed=ds_observed)
}

