% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-manifest-explore-univariate.R
\name{CreateManifestExploreUnivariate}
\alias{CreateManifestExploreUnivariate}
\title{Create a manifest for exploratoring univariate patterns.}
\usage{
CreateManifestExploreUnivariate(
    ds_observed, 
    writeToDisk = TRUE,
    pathOut = getwd(), 
    overwriteFile = FALSE,
    defaultClassGraph = c(
      numeric = "HistogramContinuous", 
      integer = "HistogramContinuous", 
      factor = "HistogramDiscrete", 
      character = "HistogramDiscrete", 
      notMatched = "histogram_generic"
    ),
    defaultFormat = c(
      numeric = "scales::comma", 
      notMatched = "scales::comma"
    ),
    bin_count_suggestion = 30L
)
}
\arguments{
\item{ds_observed}{The \code{data.frame} to create metadata for.}

\item{writeToDisk}{Indicates if the meta-dataset should be saved as a CSV.}

\item{pathOut}{The file path to save the meta-dataset.}

\item{overwriteFile}{Indicates if the CSV of the meta-dataset should be overwritten if a file already exists at the location.}

\item{defaultClassGraph}{A \code{character} array indicating which graph should be used with variables of a certain class.}

\item{defaultFormat}{A \code{character} array indicating which formatting function should be displayed on the axis of the univariate graph.}

\item{bin_count_suggestion}{An \code{integer} value of the number of roughly the number bins desired for a histogram.}
}
\value{
Returns a \code{data.frame} where each row in the metadata represents a column in \code{ds_observed}.
The metadata contains the following columns
\enumerate{
\item{\code{variable_name} The variable name (in \code{ds_observed}). \code{character}.}
\item{\code{remark} A blank field that allows theuser to enter notes in the CSV for later reference.}
\item{\code{class} The variable's \code{\link{class}} (eg, numeric, Date, factor).  \code{character}.}
\item{\code{should_graph} A boolean value indicating if the variable should be graphed. \code{logical}.}
\item{\code{graph_function} The name of the function used to graph the variable. \code{character}.}
\item{\code{x_label_format} The name of the function used to format the \emph{x}-axis. \code{character}.}
\item{\code{bin_width} The uniform width of the bins. \code{numeric}.}
\item{\code{bin_start} The location of the left boundary of the first bin. \code{numeric}.}
}
}
\description{
This function creates a meta-dataset (from the \code{data.frame} passed as a parameter) 
and optionally saves the meta-dataset as a CSV.  The meta-dataset specifies how the variables
should be plotted.
}
\examples{

CreateManifestExploreUnivariate(datasets::InsectSprays, writeToDisk=FALSE)

#Careful, the first column is a `ts` class.
CreateManifestExploreUnivariate(datasets::freeny, writeToDisk=FALSE)  
}
\keyword{explore}

